requirements.txt
pydantic
aiofiles
pydantic-ai-slim[google]
pymupdf
